+++ src/im2im/knowledge_graph_construction/edge_factories/numpy.py
@@ -1,5 +1,5 @@
 from .type import Conversion, FactoriesCluster
-from ...util import are_both_same_data_repr, is_differ_value_for_key
+from ..metedata import are_both_same_data_repr, is_differ_value_for_key
 
 
 def is_metadata_valid_for_numpy(metadata):
