+++ src/im2im/util.py
@@ -62,11 +62,3 @@ def exclude_key_from_list(keys, exclude_key):
     # New list with the exclude_key removed
     filtered_keys = [key for key in keys if key != exclude_key]
     return filtered_keys
-
-
-def are_both_same_data_repr(metadata_a, metadata_b, data_repr):
-    return metadata_a.get('data_representation') == data_repr and metadata_b.get('data_representation') == data_repr
-
-
-def is_differ_value_for_key(metadata_a, metadata_b, key):
-    return metadata_a[key] != metadata_b[key]
